Move test functions out of utils
